Políticas de acceso

+++ app/Http/Controllers/PostController.php
@@ -26,9 +26,12 @@ class PostController extends Controller
         return back()->with('status', 'Publicación guardada exitosamente');
 
     }
-    public function destroy(Post $post)
+    public function destroy(Request $request, Post $post)
     {
+        $this->authorize('delete',$post);
+
         $post->delete();
+
         return back();
     }
 }
